[hw14-spring-mvc] Moved login processing to PageLoginController
